m365-agent/src/requirements.txt
microsoft-agents-builder>=1.0.0
microsoft-agents-hosting-aiohttp>=1.0.0
azure-identity>=1.17.1
azure-search-documents>=11.5.1
azure-cosmos>=4.7.0
openai>=1.51.0
httpx>=0.27.2
pydantic>=2.9.2
